refactor(nvim): namespace change
Added the `user` namespace to the lua config for neovim.
