+++ stow/nvim/.config/nvim/lua/user/binds.lua
