+++ stow/nvim/.config/nvim/lua/user/options.lua
