+++ stow/nvim/.config/nvim/lua/user/plugins/treesitter-config.lua
